+++ pNeuma_simulator/gang/__init__.py
@@ -1,7 +1,6 @@
-# from .collision_exception import CollisionException  # noqa F401
-# from .collisions import collisions  # noqa F401
-# from .navigate import navigate  # noqa F401
-# from .neighborhood import neighborhood  # noqa F401
-# from .particle import Particle  # noqa F401
+from .collision import collisions, newton_iteration  # noqa F401
+from .navigation import navigate, egress, target, decay  # noqa F401
+from .neighborhood import neighborhood  # noqa F401
+from .particle import Particle  # noqa F401
 
 __version__ = "0.0.0"
