+++ pNeuma_simulator/initialization/__init__.py
@@ -1,6 +1,5 @@
 from .initialization import budget, f, vo, ov  # noqa F401
-from .equilibrium import equilibrium  # noqa F401
+from .equilibrium import equilibrium, synthetic_fd  # noqa F401
 from .poissondisc import PoissonDisc  # noqa F401
-from .synthetic_fd import synthetic_fd  # noqa F401
 
 __version__ = "0.0.0"
