+++ pNeuma_simulator/shadowcasting/__init__.py
@@ -1,6 +1,4 @@
-# from .fov import FoV  # noqa F401
-# from .quadrant import Quadrant  # noqa F401
-# from .row import Row  # noqa F401
-# from .shadowcasting import shadowcasting  # noqa F401
+from .fov import FoV, Row, Quadrant  # noqa F401
+from .shadowcasting import shadowcasting  # noqa F401
 
 __version__ = "0.0.0"
