--- pNeuma_simulator/shadowcasting/quadrant.py
@@ -1,20 +1,0 @@
-class Quadrant:
-    north = 0
-    east = 1
-    south = 2
-    west = 3
-
-    def __init__(self, cardinal, origin):
-        self.cardinal = cardinal
-        self.ox, self.oy = origin
-
-    def transform(self, tile):
-        row, col = tile
-        if self.cardinal == self.north:
-            return (self.ox + col, self.oy - row)
-        if self.cardinal == self.south:
-            return (self.ox + col, self.oy + row)
-        if self.cardinal == self.east:
-            return (self.ox + row, self.oy + col)
-        if self.cardinal == self.west:
-            return (self.ox - row, self.oy + col)
