+++ pNeuma_simulator/utils/__init__.py
@@ -1,10 +1,3 @@
-# from .decay import decay  # noqa F401
-# from .direction import direction  # noqa F401
-# from .egress import egress  # noqa F401
-# from .projection import projection  # noqa F401
-# from .radial import radial  # noqa F401
-# from .tangent_dist import tangent_dist  # noqa F401
-# from .truncated_rvs import truncated_rvs  # noqa F401
-# from .target import target  # noqa F401
+from .utils import direction, projection, radial, tangent_dist, truncated_rvs  # noqa F401
 
 __version__ = "0.0.0"
